+++ ChangeLog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.1] - 2020-04-19
 ### Fixed
+- Correct shading in linear color space projects due to correct (linear) sampling of normal, occlusion and metallic-roughness maps
 - Memory leak: free up volatile array `imageFormats`
 
 ## [1.0.0] - 2020-03-13